Parameterized tests in separate class, fixed method naming formant, removed unused methods, changed dependency injection
