+++ src/main/java/com/example/Animal.java
@@ -17,7 +17,4 @@ public class Animal {
     public String getFamily() {
         return "Существует несколько семейств: заячьи, беличьи, мышиные, кошачьи, псовые, медвежьи, куньи";
     }
-
-
-
 }
