+++ src/test/java/com/example/LionParameterizedTest.java
@@ -1,0 +1,39 @@
+package com.example;
+
+import org.junit.Test;
+import org.junit.runner.RunWith;
+import org.junit.runners.Parameterized;
+import org.mockito.Mock;
+import org.mockito.MockitoAnnotations;
+import static org.junit.Assert.assertEquals;
+
+
+@RunWith(Parameterized.class)
+public class LionParameterizedTest {
+
+    @Mock
+    private Feline feline;
+
+    private final String sex;
+    private final boolean expectedHasMane;
+
+    public LionParameterizedTest(String sex, boolean expectedHasMane) {
+        this.sex = sex;
+        this.expectedHasMane = expectedHasMane;
+        MockitoAnnotations.initMocks(this);
+    }
+
+    @Parameterized.Parameters
+    public static Object[][] getTestData() {
+        return new Object[][] {
+                {"Самец", true},
+                {"Самка", false}
+        };
+    }
+
+    @Test
+    public void lionHasManeTest() throws Exception {
+        Lion lion = new Lion(sex, feline);
+        assertEquals(expectedHasMane, lion.doesHaveMane());
+    }
+}
